% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alilooksee.R
\name{alilooksee}
\alias{alilooksee}
\title{A View wrapper}
\usage{
alilooksee(x, title)
}
\arguments{
\item{x}{"an R object which can be coerced to a data frame with non-zero numbers of rows and columns."}

\item{title}{"title for viewer window. Defaults to name of x."}
}
\description{
This function is a wrapper of View for those of us who like the dirtyverse. (It's an inside joke...you had to be there)
}
\examples{
alilooksee(cars)
}
\keyword{looksee}
