% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look.R
\name{look}
\alias{look}
\title{A View wrapper}
\usage{
look(x, title1)
}
\arguments{
\item{x}{"an R object which can be coerced to a data frame with non-zero numbers of rows and columns."}

\item{title}{"title for viewer window. Defaults to name of x."}
}
\description{
This function is a wrapper of View for those of us who hate capitalizing View. Quotes taken from View. See ?View for more info.
}
\examples{
look(cars)
}
\keyword{look}
